fix: Fix @val.custom not being detected on @plumier/mongoose
