+++ packages/core/src/types.ts
@@ -277,6 +277,17 @@ export interface PlumierApplication extends Application {
 // --------------------------------------------------------------------- //
 
 
+declare module "typedconverter" {
+    namespace val {
+        export function custom(validator: CustomValidatorFunction): (...arg: any[]) => void
+        export function custom(validator: CustomValidator): (...arg: any[]) => void
+        export function custom(id: string): (...arg: any[]) => void
+        export function custom(id: symbol): (...arg: any[]) => void
+        export function custom(val: CustomValidatorFunction | string | symbol | CustomValidator): (...arg: any[]) => void
+        export function result(path: string, messages: string | string[]): AsyncValidatorResult[]
+    }
+}
+
 export interface ValidatorDecorator {
     type: "ValidatorDecorator",
     validator: CustomValidatorFunction | string | symbol,
